Fix package.json command chaining syntax (#1050)
In package.json, we have several scripts which are a chain of commands.
We've been chaining them together using semicolons to separate them.
With that approach, if one command fails, it will continue and keep
running the subsequent commands, rather than failing fast. The last
command in the chain will determine the exit status of the compound
script.

That's not great, because it means that, for example, prettier
formatting is not being enforced, because prettier is not the last
command in the lint or lint-pretty chain.

This commit switches over to using `&&` to join together a chain of
commands, which will exit as soon as any link in the chain exits with a
non-zero status.

This commit also prettier-formats a few files that snuck through due to
this issue.