Add favicon to storybook
When we deploy our storybook to github pages, the URL looks something
like: https://vimeo.github.io/iris/sb/main/

The storybook build does include a favicon at
https://vimeo.github.io/iris/sb/main/favicon.ico and it's fine (it's the
storybook logo). But the generated HTML does not include a link tag
telling the browser to use that favicon, and so it falls back to trying
to load a favicon from https://vimeo.github.io/favicon.ico which
responds with a 404.

We can add this magic .storybook/manager-head.html file to inject some
code into the head of the storybook, and do whatever we like.

In this case, I'm suggesting to use the flowers emoji as an homage to
the name of the project. We could also probably use a proper image if we
want...

Snippet via https://emojicon.dev/